feat: #4 integrate with Beacon
This commit integrates the Dataset Discovery
Service with the Beacon Network.
It adds the following features:
- Call to `/2.0.0/filtering_terms` to retrieve Beacon facets;
- Call to `/2.0.0/individuals` to retrieve datasets;
- Call to `/LSAAI/token` to retrieve the LSAAI access token;

Beacon Network does not provide DCAT-AP metadata.
Instead, it provides a list of datasets and record counts
hit by the beacon query.

In that sense, `BeaconIndividualsSearchService` is the main class
of this commit. It decorates existing `DatasetsSearchService`,
adding the call to `/2.0.0/individuals`. It passes to
`DatasetsSearchService` the list of dataset identifiers as CKAN facets.

It also enhances search facets, to include beacon filtering terms,
fetched from `/2.0.0/filtering_terms`, by `BeaconFilteringTermsService`.

`facetGroupCount` is a new field in `DatasetsSearchResponse`. It
is the count per facet group, either `ckan` or `beacon`. This is
important to find any descrepancies between the number of datasets
returned by CKAN and Beacon (e.g. if CKAN returns 0 but Beacon
returns 0>, it is a sign that CKAN filters were incorrectly selected).