fix: #17 Retrieve Beacon Facets when logged-in
Beacon facets must be retrieved if the user is logged in, no matter if the user is using or not beacon facets to filter the results.

The class `BeaconDatasetsSearchService` was altered to comply to this requirement:
- if the user is not logged-in automatically CKAN query is triggered without enhancement. Check line 75.
- If there are beacon queries we trigger a search on Beacon Network. Check lines 79 and 109.
- if there are no beacon filters or beacon's resultsets is not empty, we query on CKAN. Check lines 81 and 134.
- Beacon facets/filters are always inserted into CKAN output. Check lines 87 and 175.

With those rules, we are able to properly enrich CKAN output with beacon response, including facets and resultsets, if there is any.

Several different test cases were introduced to cover most of the scenarios.